requirements.txt
aiogram>=3.0.0,<4.0.0
aioredis>=2.0.0
loguru>=0.6.0
sentry-sdk>=1.23.0
SQLAlchemy>=2.0.0
asyncpg>=0.27.0
openai>=0.27.8
cryptography>=41.0.2
requests>=2.31.0
yadisk>=1.14.0
aio-pika==6.8.0
asyncpg==0.27.0
redis==4.5.1
python-dotenv==1.0.0
pycryptodome==3.17
aiohttp==3.8.4
bcrypt==4.0.1